+++ src/addons/display_commander/config/override_reshade_settings.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace reshade::api {
+class effect_runtime;
+}
+
+void OverrideReShadeSettings(reshade::api::effect_runtime* runtime = nullptr);