+++ src/addons/display_commander/reshade_event_registration.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <windows.h>
+
+void RegisterReShadeEvents(HMODULE h_module);